third-party: add a python script to run build.rs
Summary:
Neither awk, sed nor bash are present on Windows, which makes it hard to build third
party crate with mode/win. Python is though, so let's replace all of the above
with a Python script.

Unfortunately, this doesn't allow Windows to build third-party crate just yet.
Somehow the build script can be run in the VS CMD, but not in Buck's
environment. I suspect that the PATH needs to be tweaked a bit.

Reviewed By: dtolnay

Differential Revision: D23373933

fbshipit-source-id: 124b4a681a82509501b7ef8e7b3100dafd00102d