reindeer: allow fixups to target specific versions
Summary:
num-bigint-0.3 has a buildscript which generates a source file, whereas 0.2
does not. A single fixup.toml will not apply to both, so this adds the ability
to make fixup configs apply to specific versions.

Currently the only way to have mulitiple configs in one fixup.toml is with
platform configs, so you can have an "applies to all" platform - `cfg(all())`.

Reviewed By: dtolnay

Differential Revision: D25579206

fbshipit-source-id: 01f6f85cf4f3a964e330a0192233834d9dccc88a