Disregard Cargo.toml repository entry for upstream_address metadata
Summary:
Where the published crate claims its source code is maintained/developed is considered irrelevant for METADATA.bzl as defined today, as far as I understand. The vendored source code in fbsource//third-party/rust is fetched from crates.io, not from any particular commit in the crate's repo.

{F487170876}

This diff just deletes any reliance on the `repository` metadata. If the METADATA.bzl spec is updated to accommodate a maintenance URL or issue tracker, we can add it back as that.

Reviewed By: jsgf

Differential Revision: D26962192

fbshipit-source-id: 3eb0501f96e962fb92bbbfce09f971f5518f786c