Optimize buckify writing steps
Summary:
All of these steps would really benefit from having the writer wrapped
in a `BufWriter`. But instead of doing that, we buffer the output so we
can see if a write is even necessary at all. Not only is this faster
than before but also avoids triggering Watchman (and Buck cache
invalidation) for every generated file.

Reviewed By: dtolnay

Differential Revision: D35757044

fbshipit-source-id: 33accf8035091e2582f79bb7b16da0b9454bf810