Make precise_srcs overridable by fixups
Summary:
When combined with D39932021 (setting `precise_srcs = false` for the
`windows` create), `buckify` then takes 52s instead of 1m47s.

About ~30s of the 52s is `cargo metadata` so there's not much left to
improve here. Though another possible optimization is to skip work for
benchmarks, tests and examples. But this only ends up saving ~2s, and it
would be pretty kludgy in the current code:

  diff --git a/fbcode/common/rust/tools/reindeer/src/buckify.rs b/fbcode/common/rust/tools/reindeer/src/buckify.rs
  --- a/fbcode/common/rust/tools/reindeer/src/buckify.rs
  +++ b/fbcode/common/rust/tools/reindeer/src/buckify.rs
  @@ -183,6 +183,11 @@

       log::info!("Generating rules for package {} target {}", pkg, tgt.name);

  +    if tgt.kind_bench() || tgt.kind_test() || tgt.kind_example() {
  +        log::debug!("pkg {} target {} skipping {:?}", pkg, tgt.name, tgt.kind());
  +        return Ok((vec![], vec![]));
  +    }
  +
       let fixups = Fixups::new(config, paths, index, pkg, tgt)?;

       if fixups.omit_target() {

Reviewed By: diliop

Differential Revision: D39932017

fbshipit-source-id: 29d9b16b8f5f44ff103ba9e715d2511112b2d181