Add workspace directive
Summary:
If there is a Cargo.toml representing a workspace higher up in the directory, then Cargo will really want this file to be a part of it. But this is a fake Cargo.toml, so that doesn't make sense. Adding `[workspace]` tells Cargo to not worry.

Not relevant in the reindeer repo (as there is no workspace), but if you are trying to Buckify a large project it can matter.

Reviewed By: zertosh

Differential Revision: D40139001

fbshipit-source-id: 0b16894b24c7378388c3cce952e93d4b0b5ab512