Visibility stable BUCK layout
Summary:
This code all goes away in D40482901. This diff is just to sort
`third-party/rust/BUCK` in the same manner as D40482901, so the actual
changes to the rules are clearly visible without the noise of rules
jumping around.

Currently, **public** crates' name is the crate name (`NAME`), and
**private** crate names is `NAME-VERSION`. When a crate changes
visibility, the rule order changes, even if the version is the same.
This is an annoyance that makes diff review harder. A side effect of
D40482901, is an ordering exclusively based on versioning - regardless
of visibility.

Reviewed By: dtolnay

Differential Revision: D40482903

fbshipit-source-id: 37ccb174e6b45e8041c900d294e150517072a38a