Serialize 'platform' in a way that is meaningful to buildifier
Summary:
Buildifier only touches function calls using `f(` `)` syntax, not stuff in `{` `}` syntax. It uses some [tables](https://github.com/bazelbuild/buildtools/blob/master/tables/tables.go) to control what named function arguments are safe to handle in what way (like you probably want it to sort your `deps` and `srcs`, but not your `rustc_flags`...) and none of that kicks in when dealing with a `{` `}`.

This diff changes reindeer to emit `platform` entries in a way that buildifier will operate on.

**Before:**

```
     platform = {
-        "android-arm64": {
-            "srcs": [
                 "vendor/bitvec-1.0.0/README.md",
                 "vendor/bitvec-1.0.0/doc/README.md",
                 "vendor/bitvec-1.0.0/doc/access/BitAccess.md",
                 "vendor/bitvec-1.0.0/doc/access/BitSafe.md",
                 "vendor/bitvec-1.0.0/doc/access/impl_BitSafe.md",
-                "vendor/bitvec-1.0.0/doc/access.md",
                 "vendor/bitvec-1.0.0/doc/array/BitArray.md",
     ...
```

**After:**

```
     platform = {
+        "android-arm64": dict(
+            srcs = [
                 "vendor/bitvec-1.0.0/README.md",
                 "vendor/bitvec-1.0.0/doc/README.md",
+                "vendor/bitvec-1.0.0/doc/access.md",
                 "vendor/bitvec-1.0.0/doc/access/BitAccess.md",
                 "vendor/bitvec-1.0.0/doc/access/BitSafe.md",
                 "vendor/bitvec-1.0.0/doc/access/impl_BitSafe.md",
+                "vendor/bitvec-1.0.0/doc/array.md",
                 "vendor/bitvec-1.0.0/doc/array/BitArray.md",
     ...
```

Notice that buildifier is now able to use its preferred order for `access.md` relative to `access/`, which is different from what Rust's `Ord` for `PathBuf` was doing.

Reviewed By: zertosh

Differential Revision: D41820075

fbshipit-source-id: 2fa0340ee6db984a01b21090a704a17392c07cd3