Eliminate RuleTarget enum
Summary:
I noticed some disturbing things about code involving `RuleTarget` in reindeer.

## 1. RuleTarget::Local may or may not contain the leading colon

The `Display` impl expects there to be **no colon** in `Local` because it adds its own.

https://www.internalfb.com/code/fbsource/[a1c22bb712a7119a96171b906ed04721108f790e]/fbcode/common/rust/tools/reindeer/src/buck.rs?lines=104-111

However, the `Deserialize` impl **includes the colon** when instantiating a `Local`.

https://www.internalfb.com/code/fbsource/[a1c22bb712a7119a96171b906ed04721108f790e]/fbcode/common/rust/tools/reindeer/src/buck.rs?lines=119-133

The wrong sequence of operations would end up with a double colon.

## 2. RuleTarget::Abs can contain local (non-absolute) targets

In fixups, `extra_deps` are *always* treated using `Abs`.

https://www.internalfb.com/code/fbsource/[a1c22bb712a7119a96171b906ed04721108f790e]/fbcode/common/rust/tools/reindeer/src/fixups.rs?lines=701-705

That's the case even if the extra dep is a local target.

https://www.internalfb.com/code/fbsource/[83672381a3912cc4eda69f7768a9d2bb160c4720]/third-party/rust/fixups/tokio/fixups.toml?lines=8

## 3. The sort order is jarring

The `PartialOrd` is derived, which orders values based on variant order.

https://www.internalfb.com/code/fbsource/[d7c4b26282f22864fc44ee017297d866c87dadb1]/fbcode/common/rust/tools/reindeer/src/buck.rs?lines=40-44

As a consequence of #2, reindeer (with buildifier off) can end up emitting "sorted" deps arrays that look like:

```
deps = [
    ":something",  # Local
    "//third-party/thing:thing",  # Abs
    ":tracing",  # Abs
]
```

 ---

Overall, `RuleTarget` has not served as a useful abstraction.

I do think there is room for an abstraction that helps distinguish local vs arbitrary targets in a way that is useful for correctness, and I'll be introducing that in the rest of this stack.

Reviewed By: zertosh

Differential Revision: D41844283

fbshipit-source-id: 970d3c6f36bf295e74cbe4079920bedc7130759f