Use Name for the target of an Alias
Summary:
Reindeer-generated `alias` targets only ever point to another target within the same Buck package. Code in `rule_sort_key` assumes that to be the case (panicking if it isn't).

This diff replaces `alias` from `RuleRef` (which can be non-local) to `Name` to enforce in the type system that all the necessary code will be modified if we ever generalize to non-local aliases.

Reviewed By: zertosh

Differential Revision: D41844282

fbshipit-source-id: a1fe387a22249dd9c224dd58b08f015ffb84fc9c