Change visibility from bool to enum
Summary:
As part of splitting third-party/rust/BUCK into one small targets file per imported crate, we are going to want visibilities like `visibility = ["//third-party/rust"]`. The plan is for the top-level BUCK file to hold only aliases, and only to those crates that are first-level dependencies of our Cargo.toml. All other crates will not be able to be referred to from outside the third-party/rust directory.

Implementing this now because it slightly simplifies D41855277.

Reviewed By: zertosh

Differential Revision: D41855276

fbshipit-source-id: 8a374f7d8effd1da83475ef536c7d79ed09fe5a0