Fix wrong cargo used due to unset config_path
Summary:
`config_path` is not read from the reindeer.toml config file (that's what `serde(skip)` means) but is passed around together with everything read from the config file so that relative paths found in the config can be interpreted correctly.

https://www.internalfb.com/code/fbsource/[d89f7112b6f83f841051d6013cda137ac6f1d492]/fbcode/common/rust/tools/reindeer/src/config.rs?lines=32-34

Sadly it was only getting filled in on one of the two necessary codepaths.

- When reindeer.toml was not found, it was filled in (but pretty useless in this case because then there aren't relative paths in the config file to interpret based on it).

- When reindeer.toml was found and deserialized, `config_path` would inadvertently get left as its default value of empty PathBuf.

Later `config_path` gets used here:

https://www.internalfb.com/code/fbsource/[d89f7112b6f83f841051d6013cda137ac6f1d492]/fbcode/common/rust/tools/reindeer/src/cargo.rs?lines=86-90

to interpret the relative path in `cargo`.

https://www.internalfb.com/code/fbsource/[d89f7112b6f83f841051d6013cda137ac6f1d492]/fbcode/common/rust/tools/reindeer/src/config.rs?lines=113-115

This accidentally used to work because we'd only ever run cargo from the same directory that the config was found in! But in D42203704 I'm adding a cargo invocation from a different directory and do need the correct `config_path`.

Reviewed By: zertosh

Differential Revision: D42203902

fbshipit-source-id: 6db17d7fc2883e823364dce08eda4f572a373de2