Support buckify without an existing .cargo/config
Summary:
Previously `reindeer buckify` would fail with an error if `.cargo/config` was not present. That file is created by `reindeer vendor` based on the output printed by `cargo vendor`.

```
$ reindeer buckify
[ERROR reindeer] parsing metadata

Caused by:
  0: running cargo
  1: /data/users/$user/fbsource/third-party/rust/../../xplat/rust/toolchain/current/cargo metadata --format-version 1 --frozen --locked --offline --manifest-path /data/users/$user/fbsource/third-party/rust/Cargo.toml failed:
     error: failed to load source for dependency `aes-gcm-siv`

     Caused by:
       Unable to update https://github.com/lei2022/AEADs?rev=d9d03399fbc1347a1c2941801c0743e847980757#d9d03399

     Caused by:
       can't checkout from 'https://github.com/lei2022/AEADs': you are in the offline mode (--offline)
```

I am interested in making this work for 3 reasons:

1. In CI of https://github.com/dtolnay/cxx, I am interested in running `reindeer buckify` without `reindeer vendor`. I just want to confirm that the checked-in generated BUCK file is always up to date.

2. In particular I do not want to run `reindeer vendor` because it is slow. The way it uses `.cargo` as the CARGO_HOME forces Cargo to clone a whole new crates.io index and all your git dependencies and download all the crate tarballs that might already be in your normal Cargo directory in your home directory. In fbsource, this currently amounts to 4.9 GB that it writes to `.cargo` on every invocation, and a bunch of network traffic. This diff is a step toward fixing `vendor` to not work this way.

3. I want to experiment with *"just delete third-party/rust/.cargo"* as the magic solution to all `git`-related `cargo vendor`-related headaches.

The content of `.cargo/config` is 100% redundant with `Cargo.lock` and can be generated from `Cargo.lock`, which is what this diff does. My suspicion is that in general, workflows that involve Cargo changing `Cargo.lock` are way more robust in Cargo than workflows that involve Cargo changing `.cargo/config`, especially when it comes to git dependencies. We'll be better off moving off of reliance on a checked-in `.cargo/config`, and instead only generating an ephemeral one from `Cargo.lock` when needed.

Reviewed By: zertosh

Differential Revision: D42203704

fbshipit-source-id: 3dd5f422e7e21e43b671eb397e490c0ce094ed9b