Allow custom visibility via fixup for public packages
Summary:
It came up in D42891044 that a crate needs to be added so it exists in
`third-party/rust/vendor` (for cargo offline use), but should not really
ever be referenced from buck. This diff adds a `visibility` config to
control the visibility of **public** packages. This is actually useful
beyond just turning `"PUBLIC"` off, but can serve to guide deprecations
by limiting a crate's visibility to those that are already using it, and
not let it spread.

Reviewed By: dtolnay

Differential Revision: D42949597

fbshipit-source-id: 15584ee280b6fe6998562c58504d4d4cfbb32cf5