Add support for CFLAGS-style env vars which contain paths
Summary:
I need this for passing `BINDGEN_EXTRA_CLANG_ARGS="-I $(location //path/to/libvirt:include)"` to virt-sys in D43266380.

Reindeer's existing `env` fixup does not work for this because the generated genrule's directory is not the directory that fbsource's implementation of buildscript_genrule_srcs uses as the current directory while running build scripts, so the relative path interpolated by Buck for `$(location …)` needs to get appropriately adjusted.

Normally Reindeer's existing `path_env` fixup does that adjusting but it does not work here because we also need an `-I` to go in front of the path part.

The approach implemented here is the easiest generalization that works for things like `CFLAGS`, `CXXFLAGS`, `RUSTFLAGS`, `RUSTDOCFLAGS`, `MIRIFLAGS`, and `BINDGEN_EXTRA_CLANG_ARGS`. We could consider doing something more sophisticated with arrays of strings (instead of splitting on space) if it becomes necessary.

Reviewed By: zertosh

Differential Revision: D43266378

fbshipit-source-id: ec0560b02eaca3c97179fb1a6bce946b833d022b