Allow setting a different macro name for buildscript binaries than other binaries
Summary:
Reindeer currently generates a publicly visible `rust_binary` for *every* binary crate inside every package that we directly depend on. For example silly things like:

https://www.internalfb.com/code/fbsource/[5e67bdcd361569e86c207f76df42a351ef1fb8ad]/third-party/rust/vendor/clap-4.1.11/src/bin/stdio-fixture.rs?lines=1-15

https://www.internalfb.com/code/fbsource/[5e67bdcd361569e86c207f76df42a351ef1fb8ad]/third-party/rust/BUCK?lines=17313-17321%2C17323-17324%2C17475

This is obviously not meant to be a public binary. It's used internally by the crate's test suite. The crate maintainer doesn't worry about it being in the package because Cargo does not provide any way for a downstream library to depend on it.

I am interested in moving all of the `rust_binary` targets we actually rely on into artifact dependencies (https://fb.workplace.com/groups/rust.language/posts/24198644003090887). Once that is done I will make a reindeer change to disable generating `rust_binary` for binaries like this `stdio-fixture` in clap which we do not declare a dependency on via an artifact dependency.

One obstacle to this is that `rg '^rust_binary(`' in third-party/rust/BUCK is unhelpful in distinguishing which crates have binaries, because it is swamped by a much larger number of crates' build scripts, which are also built by a `rust_binary` call.

This diff adds a way to make reindeer use 2 different macros for binary crates (in the future: only artifact dependencies) vs for build scripts. By default, build scripts will use the same macro that the configuration says binaries should use, but can be configured to use a different macro.

Reviewed By: zertosh

Differential Revision: D44565271

fbshipit-source-id: 0cc999f6a5a3b44de7abbe386a2f76a0063f0cbb