Delete support for extra_top_levels
Summary: The purpose of this configuration was to "include targets for binary and cdylib-only packages (that is packages which can't directly be depended-on by Rust)". However Cargo can now depend on such packages using an artifact dependency, and extra_top_levels has been superseded by reindeer support for artifact dependencies.

Reviewed By: zertosh

Differential Revision: D44661586

fbshipit-source-id: 3260297462b9b5ac2728c72573d29d73004cc62d