Traverse cargo's selected target kinds only
Summary:
Previously, for every dependency edge from package A on package B, reindeer would go generate all of the targets contained in package B.

That approach is incompatible with supporting non-top-level artifact dependencies, i.e. when rust-third-party depends on some library which depends on some artifact. That's because reindeer would have no way to know whether to generate targets for B's bins. It was throwing away information about which piece of package B was being depended on by package A.

The initial support for artifact dependencies worked around this by simply assuming non-top-level artifact dependencies did not exist. Under that assumption, we can emit bins for crate B if and only if B is top-level.

This diff updates reindeer to keep track of which piece of a package is required by the incoming dependency edge. If something is depending on bins, we can generate targets for the bins; if something is depending on the library / proc-macro, we can generate the target for that.

Reviewed By: zertosh

Differential Revision: D44860684

fbshipit-source-id: 781b5c1a60cd05cb937bd71bde350abd020ce193