Improve validation of visibility fixup
Summary:
Reindeer wants to prevent you from writing a `visibility` fixup on something that is only a transitive dependency, i.e. a package where none of the targets contained in it are direct dependencies of the top-level Cargo.toml.

This check was previously implemented using:

```
!target
    .kind
    .iter()
    .any(|kind| index.is_public(package, *kind))
```

but that is incompatible with having a different value of `is_public` for different binaries inside the same package.

This diff implements the check in a better way to unblock the next diff in this stack (D44860683).

Reviewed By: zertosh

Differential Revision: D44860682

fbshipit-source-id: 0543e452c46fad790a82086ce251d2b6092155bf