Be precise about which bins are being depended on
Summary:
Previously if something depended on any binary from a particular package, reindeer would generate targets for *all* of that package's binaries.

This diff updates reindeer to keep track of specifically which binary from a package is being depended on, and only generate a target for the binaries for which one is needed.

Reviewed By: zertosh

Differential Revision: D44860683

fbshipit-source-id: 847e57fa55b8032052cb507be5b3d2039ff21a09