Drive top level of buckify using only root pkg, not all public pkgs
Summary:
Currently reindeer buckify's outer loop looks like this:

```
for crate in public_crates:
    transitively_generate_targets(crate)
```

This means that if you have a crate that you want to vendor but not buckify, you need to express that *twice*:

- Once in fixups/rust-third-party/fixups.toml using `omit_deps`, to keep the top-level crate (in `include_top_level = true` mode) from listing the non-buck crate in its `deps`

- Once in fixups/nonbuck-crate/fixpus.toml using `omit_targets` to block a target from being generated for the crate.

If you only include `omit_deps` without `omit_targets`, a library target is still generated for the crate that you didn't want a buck target for.

If you only include `omit_targets` without `omit_deps`, reindeer generates broken targets because the root crate will get a dependency on a target which doesn't exist.

Ideally only one of these would be sufficient. The one that makes sense to me to keep is the `omit_deps`. This diff changes reindeer buckify's outer loop to:

```
transitively_generate_targets(root)
```

while suppressing a target corresponding to the root crate if `include_top_level` is configured to false.

Reviewed By: zertosh

Differential Revision: D45430862

fbshipit-source-id: 1ca691e0bb68f093e77dac2d472db2f726ff1c6f