Do not recurse into omitted deps
Summary:
While processing the dependency graph during buckify, if we land on a crate that has an unconditional `omit_deps = ["some_dependency"]` i.e. not a platform-specific omit, then do not traverse into generating targets for that dependency.

It will still get targets generated if there is something *else* which depends on the same dependency and does not also omit it.

Reviewed By: zertosh

Differential Revision: D45430865

fbshipit-source-id: 9c5b3d1f059cf38afa3ab5fce860dfed31be31a8