Delete support for buildscript.build fixup
Summary:
`[buildscript.build]` makes reindeer generate a rust_binary target for your build script, but not wire it up to anything (such as OUT_DIR or rustc_flags).

As far as I can tell, the intended role of this fixup was that you would put this, buckify, `buck run` the build script to observe what happens, and then replace it with the actual correct fixup based on what you learned. It is never something that you are supposed to check in.

Despite that, *lots* of people end up checking it in. See {D45430860} and D45436276.

In my opinion most people would not find it valuable to `buck run` a build script. The few people who occasionally want that would know to write `[buildscript.rustc_flags]` or `[buildscript.gen_srcs]` instead and obtain a target to run that way.

Reviewed By: zertosh

Differential Revision: D45448075

fbshipit-source-id: 3e1b85a2be57457f0ef3321244205b9b927aeb15