Own implementation of srcfiles
Summary:
This is a re-write of [srcfiles] in a way that is more tailored to what
Reindeer needs. This is a much simpler and shorter implementation -
although it doesn't look like it because of all the tests.

This uses syn1 because there are a few files that no longer parse with
syn2. The next diff switches to syn2. It's not done here so it's easier
to compare the old srcfiles against the new srcfiles.

[srcfiles]: https://github.com/jsgf/srcfiles

Reviewed By: dtolnay

Differential Revision: D45437068

fbshipit-source-id: f220ce166b008cede3369ba8c35eae6a654a0912