Adjust how default rule names are captured in configuration
Summary:
In {D45498484}, I am trying to get reindeer to produce working targets out of the box with minimal config.

I want it to fill in an appropriate working default value for `buckfile_imports`, `rust_library`, `rust_binary`, and `buildscript_genrule`.

However, just defaulting all of those fields on an individual basis is not good. If someone configured `buckfile_imports` but not `rust_library`, it's not safe to assume they still want `rust_library` to try to refer to the prelude's rust_library. Inversely if someone configured `rust_library` but not `buckfile_imports`, filling in a `buckfile_imports` with prelude imports behind their back is not correct -- they may not even be using the prelude that ships with buck2, and the value they set for `rust_library` may be something that's already put into scope in all modules by their local prelude, without needing any import.

This diff doesn't immediately change any behavior of reindeer, but it makes it possible for D45498484 to distinguish whether a configuration was set explicitly in the config vs assigned its default value.

Reviewed By: zertosh

Differential Revision: D45498472

fbshipit-source-id: 0bc89a91b8b29d628956236bbf1f2a36d91761da