Hide target_abi="llvm" deps from default windows-gnu platform
Summary:
Rust has relatively new (new in Rust 1.62) Tier 3 targets for `aarch64-pc-windows-gnullvm` and `x86_64-pc-windows-gnullvm`. These are niche targets compared to the Tier 1 `x86_64-pc-windows-gnu` target which has been Tier 1 for as long as the tier system has existed.

https://doc.rust-lang.org/1.69.0/rustc/platform-support.html
https://doc.rust-lang.org/1.69.0/rustc/platform-support/pc-windows-gnullvm.html

The `windows-targets` crate recently started having deps specific to `not(target_abi = "llvm")` vs `target_abi = "llvm"`:

https://www.internalfb.com/code/fbsource/[5f3ee66a7b8f12023af43eb5951cc558f29cf42a]/third-party/rust/vendor/windows-targets-0.48.0/Cargo.toml?lines=34-35%2C37-38

For the purpose of this `windows-gnu` reindeer default platform, the `not(target_abi = "llvm")` is the set we want. Previously it would apply both because the platform selection condition would be ignorant to target_abi.

Reviewed By: zertosh

Differential Revision: D45608483

fbshipit-source-id: 3666bd6142c1ff80e1b3f64117810aeac3e36047