Serialize include-directories as preprocessor flags
Summary:
This diff replaces `include_directories = ["$(location ...)/path"]` with `preprocessor_flags = ["-I$(location ...)/path"]` in generated `cxx_library` targets.

Using `preprocessor_flags = ["-I..."]` is not desirable. There is a comment in fbcode_macros that says we'd like to eliminate all such usage using a codemod, in favor of `include_directories` (as initially implemented in {D45638529}).

https://www.internalfb.com/code/fbsource/[8e8e736e14d23e77fa2bd5e72123d66943f7716f]/tools/build_defs/fbcode_macros/build_defs/lib/cpp_common.bzl?lines=397-399

However, doing this based on `preprocessor_flags` avoids the need for {D45639663} which got reverted in D45682623.

Reviewed By: zertosh

Differential Revision: D45683164

fbshipit-source-id: 558981a06c033668e777e9b5033e6c2a053d3347