Generate platform-specific deps for default platforms if no reindeer.toml
Summary:
Fixes https://github.com/facebookincubator/reindeer/issues/17.

Previously if you used an empty `reindeer.toml`, it would get deserialized using serde and the toml crate, and the following default platforms would kick in:

https://www.internalfb.com/code/fbsource/[35fb6a0192b584b38ea216dcc8ab3537dbaf19d7]/fbcode/common/rust/tools/reindeer/src/config.rs?lines=90-91

But if there was no reindeer.toml, Reindeer would use the `impl Default for Config`, which does not respect that attribute.

I don't know of a good custom `derive(Default)` library that is customizable with attributes, but if one exists, it may be appropriate to use that.

Reviewed By: zertosh

Differential Revision: D47316923

fbshipit-source-id: fffd4dd1dd3cd8d04804f22ef2a8c2c6e1785752