Upgrade flate2 to the latest version
Summary:
This diff update `flate2` crate to the current latest version.

As a result of upgrading the `flate2` crate, the `async-compression` shed library (created by Rain) broke. The library defines `RawEncoder` trait that is implemented for each encoder type (`Gzip`, `Zlib`, etc.) and requires that each underlying encoder implement the `tokio::AsyncWrite` trait. With the latest version of `flate2`, that is no longer true for `ZlibEncoder`. I applied a similar workaround to what is already being used for `ZstdEncoder` by creating an `Async` wrapper around the encoder type which implements the `tokio::AsyncWrite` trait. The unit tests from the crate validate that encoding and decoding works as expected.

Reviewed By: zertosh

Differential Revision: D47913781

fbshipit-source-id: 9281b531d1e3d2fce3ae9e3b7765e05dc00e5340