Fix install instructions (#24)
Summary:
The install instructions in the README fail with

```
error: multiple packages with binaries found: reindeer, rust-third-party. When installing a git repository, cargo will always search the entire repo for any Cargo.toml.
```

This just adds `reindeer` to the install so the first time experience is better.

Pull Request resolved: https://github.com/facebookincubator/reindeer/pull/24

Reviewed By: zertosh

Differential Revision: D48433577

Pulled By: shayne-fletcher

fbshipit-source-id: 6a61acaae789737da96146b3789cb3ca1346feee