Remove pointless comparison (#26)
Summary:
When parsing the cargo sources the code was comparing the rev specified by the user to the full sha-1 commit hash that cargo locked the git repository to. This would cause reindeer buckify to fail if the user specified a short (7+) hash, a common practice. This check is really rather pointless, as the commit hash is the only thing that truly matters, and if the user did in fact specify an ambiguous/incorrect hash, short or otherwise, cargo would fail to get the correct full sha-1 commit hash anyways, something that reindeer itself need not handle, nor care about, at least in this context.

Pull Request resolved: https://github.com/facebookincubator/reindeer/pull/26

Reviewed By: zertosh

Differential Revision: D48433699

Pulled By: shayne-fletcher

fbshipit-source-id: 51ffea829d82ca44d6cee1b6ad00b5b174a1bf27